moving typography into components
