+++ assets/src/design-system/components/index.js
@@ -14,5 +14,4 @@
  * limitations under the License.
  */
 
-export * from './theme';
-export * from './components';
+export * from './typography';
