+++ assets/src/design-system/components/typography/display/index.js
@@ -23,35 +23,26 @@ import PropTypes from 'prop-types';
 /**
  * Internal dependencies
  */
-import { THEME_CONSTANTS } from '../../theme';
+import { THEME_CONSTANTS } from '../../../';
 import { defaultTypographyStyle } from '../styles';
 
-export const Text = styled.p`
+export const Display = styled.h1`
   ${defaultTypographyStyle}
-  ${({ as, isBold, size, theme }) => {
-    const { text } = theme.typography.presets;
-    const asLink = as === 'a' && {
-      color: theme.colors.accent.secondary,
-      textDecoration: 'none',
-      cursor: 'pointer',
-    };
+  ${({ size, theme }) => {
+    const { display } = theme.typography.presets;
     return {
-      fontSize: `${text[size].size}px`,
-      fontWeight: isBold ? theme.typography.weight.bold : text[size].weight,
-      lineHeight: `${text[size].lineHeight}px`,
-      letterSpacing: `${text[size].letterSpacing}em`,
-      ...asLink,
+      fontSize: `${display[size].size}px`,
+      fontWeight: display[size].weight,
+      lineHeight: `${display[size].lineHeight}px`,
+      letterSpacing: `${display[size].letterSpacing}px`,
     };
   }}
 `;
 
-Text.propTypes = {
-  as: PropTypes.oneOf(['p', 'a', 'span']),
-  isBold: PropTypes.bool,
-  size: PropTypes.oneOf(THEME_CONSTANTS.TEXT_SIZES),
+Display.propTypes = {
+  as: PropTypes.oneOf(['h1', 'h2', 'h3', 'h4', 'h5', 'h6']).isRequired,
+  size: PropTypes.oneOf(THEME_CONSTANTS.DISPLAY_SIZES),
 };
-Text.defaultProps = {
-  as: 'p',
-  isBold: false,
+Display.defaultProps = {
   size: THEME_CONSTANTS.TYPOGRAPHY_PRESET_SIZES.MEDIUM,
 };
